+++ src/CapabilityService.Tests/Builders/CapabilityApplicationServiceBuilder.cs
@@ -1,6 +1,7 @@
 using DFDS.CapabilityService.Tests.TestDoubles;
 using DFDS.CapabilityService.WebApi.Application;
 using DFDS.CapabilityService.WebApi.Domain.Repositories;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Repositories;
 
 namespace DFDS.CapabilityService.Tests.Builders
 {
