+++ src/CapabilityService.Tests/Builders/MessageContractBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using DFDS.CapabilityService.WebApi.Domain.Models;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Models;
 
 namespace DFDS.CapabilityService.Tests.Builders
 {
