+++ src/CapabilityService.Tests/TestDoubles/StubTopicRepository.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using DFDS.CapabilityService.WebApi.Domain.Models;
 using DFDS.CapabilityService.WebApi.Domain.Repositories;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Models;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Repositories;
 
 namespace DFDS.CapabilityService.Tests.TestDoubles
 {
