+++ src/CapabilityService.WebApi/Application/CapabilityApplicationService.cs
@@ -5,6 +5,9 @@ using System.Threading.Tasks;
 using DFDS.CapabilityService.WebApi.Domain.Exceptions;
 using DFDS.CapabilityService.WebApi.Domain.Models;
 using DFDS.CapabilityService.WebApi.Domain.Repositories;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Exceptions;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Models;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Repositories;
 
 namespace DFDS.CapabilityService.WebApi.Application
 {
