--- src/CapabilityService.WebApi/Application/ITopicApplicationService.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace DFDS.CapabilityService.WebApi.Application
-{
-    public interface ITopicApplicationService
-    {
-        Task UpdateMessageContract(Guid topicId, string type, string description, string content);
-        Task RemoveMessageContract(Guid topicId, string type);
-        Task UpdateTopic(Guid topicId, string name, string description, bool isPrivate);
-    }
-}
