+++ src/CapabilityService.WebApi/CapabilityService.WebApi.csproj
@@ -31,6 +31,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Features\Topics\Infrastructure" />
     <Folder Include="Properties" />
   </ItemGroup>
