--- src/CapabilityService.WebApi/Domain/Exceptions/TopicAlreadyExistException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace DFDS.CapabilityService.WebApi.Domain.Exceptions
-{
-    public class TopicAlreadyExistException : Exception
-    {
-        public TopicAlreadyExistException(string message) : base(message)
-        {
-            
-        }
-    }
-}
