--- src/CapabilityService.WebApi/Domain/Exceptions/TopicDoesNotExistException.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace DFDS.CapabilityService.WebApi.Domain.Exceptions
-{
-    public class TopicDoesNotExistException : Exception
-    {
-        public TopicDoesNotExistException(Guid topicId) 
-            : base($"Topic with id \"{topicId}\" does not exist.")
-        {
-            
-        }
-    }
-}
