+++ src/CapabilityService.WebApi/Features/Topics/Application/ITopicApplicationService.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace DFDS.CapabilityService.WebApi.Features.Topics.Application
+{
+    public interface ITopicApplicationService
+    {
+        Task UpdateMessageContract(Guid topicId, string type, string description, string content);
+        Task RemoveMessageContract(Guid topicId, string type);
+        Task UpdateTopic(Guid topicId, string name, string description, bool isPrivate);
+    }
+}
