+++ src/CapabilityService.WebApi/Features/Topics/Domain/Exceptions/TopicAlreadyExistException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace DFDS.CapabilityService.WebApi.Features.Topics.Domain.Exceptions
+{
+    public class TopicAlreadyExistException : Exception
+    {
+        public TopicAlreadyExistException(string message) : base(message)
+        {
+            
+        }
+    }
+}
