+++ src/CapabilityService.WebApi/Features/Topics/Domain/Exceptions/TopicDoesNotExistException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace DFDS.CapabilityService.WebApi.Features.Topics.Domain.Exceptions
+{
+    public class TopicDoesNotExistException : Exception
+    {
+        public TopicDoesNotExistException(Guid topicId) 
+            : base($"Topic with id \"{topicId}\" does not exist.")
+        {
+            
+        }
+    }
+}
