+++ src/CapabilityService.WebApi/Features/Topics/Infrastructure/Api/DTOs/MessageContract.cs
@@ -1,0 +1,19 @@
+﻿namespace DFDS.CapabilityService.WebApi.Features.Topics.Infrastructure.Api.DTOs
+{
+    public class MessageContract
+    {
+        public string Type { get; set; }
+        public string Content { get; set; }
+        public string Description { get; set; }
+
+        public static MessageContract CreateFrom(Domain.Models.MessageContract messageContract)
+        {
+            return new MessageContract
+            {
+                Type = messageContract.Type,
+                Content = messageContract.Content,
+                Description = messageContract.Description
+            };
+        }
+    }
+}
