+++ src/CapabilityService.WebApi/Features/Topics/Infrastructure/Api/DTOs/MessageContractInput.cs
@@ -1,0 +1,8 @@
+﻿namespace DFDS.CapabilityService.WebApi.Features.Topics.Infrastructure.Api.DTOs
+{
+    public class MessageContractInput
+    {
+        public string Description { get; set; }
+        public string Content { get; set; }
+    }
+}
