+++ src/CapabilityService.WebApi/Features/Topics/Persistence/TopicTransactionalDecorator.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Threading.Tasks;
 using DFDS.CapabilityService.WebApi.Application;
+using DFDS.CapabilityService.WebApi.Features.Topics.Application;
+using DFDS.CapabilityService.WebApi.Infrastructure.Persistence;
 
-namespace DFDS.CapabilityService.WebApi.Infrastructure.Persistence
+namespace DFDS.CapabilityService.WebApi.Features.Topics.Persistence
 {
     public class TopicTransactionalDecorator : TransactionalDecoratorBase, ITopicApplicationService
     {
