--- src/CapabilityService.WebApi/Infrastructure/Api/DTOs/MessageContractInput.cs
@@ -1,8 +1,0 @@
-﻿namespace DFDS.CapabilityService.WebApi.Infrastructure.Api.DTOs
-{
-    public class MessageContractInput
-    {
-        public string Description { get; set; }
-        public string Content { get; set; }
-    }
-}
