+++ src/CapabilityService.WebApi/Startup.cs
@@ -15,6 +15,9 @@ using DFDS.CapabilityService.WebApi.Enablers.CorrelationId;
 using DFDS.CapabilityService.WebApi.Enablers.KafkaStreaming;
 using DFDS.CapabilityService.WebApi.Enablers.Metrics;
 using DFDS.CapabilityService.WebApi.Enablers.PrometheusHealthCheck;
+using DFDS.CapabilityService.WebApi.Features.Topics.Application;
+using DFDS.CapabilityService.WebApi.Features.Topics.Domain.Repositories;
+using DFDS.CapabilityService.WebApi.Features.Topics.Persistence;
 using DFDS.CapabilityService.WebApi.Infrastructure.Events;
 using DFDS.CapabilityService.WebApi.Infrastructure.Messaging;
 using DFDS.CapabilityService.WebApi.Infrastructure.Persistence;
